requirements.txt
absl-py==2.1.0
adagio==0.2.4
antlr4-python3-runtime==4.11.1
appdirs==1.4.4
asgiref==3.6.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
azure-core==1.26.2
azure-storage-blob==12.14.1
bleach==5.0.1
boto3==1.16.63
botocore==1.19.63
Brotli==1.1.0
cachetools==5.3.0
catboost==1.2.5
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==3.0.0
cmdstanpy==1.2.4
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
convertdate==2.4.0
coreapi==2.3.3
coreforecast==0.0.10
coreschema==0.0.4
cryptography==39.0.0
cycler==0.12.1
Cython==3.0.10
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.16
django-annoying==0.10.6
django-cors-headers==3.6.0
django-debug-toolbar==3.2.1
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-rest-swagger==2.2.0
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
drf-generators==0.3.0
drf-yasg==1.20.0
ephem==4.1.5
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
fonttools==4.53.1
fs==2.4.16
fsspec==2024.6.1
fugue==0.9.1
gast==0.4.0
google-api-core==2.10.0
google-auth==2.26.2
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.3.2
google-cloud-logging==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
graphviz==0.20.3
grpc-google-iam-v1==0.12.3
grpcio==1.51.1
grpcio-status==1.48.2
h2==4.1.0
h5py==3.10.0
holidays==0.53
hpack==4.0.0
humanize==4.10.0
hyperframe==6.0.1
idna==3.4
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflection==0.5.1
ipykernel==6.29.5
ipython==8.26.0
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jsonschema==3.2.0
jupyter_client==8.6.2
jupyter_core==5.7.2
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
label-studio==1.7.1
label-studio-converter==0.0.48rc0
label-studio-tools==0.0.1
launchdarkly-server-sdk==7.5.0
lightgbm==4.4.0
lightning-utilities==0.11.5
llvmlite==0.42.0
LunarCalendar==0.0.9
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.9.1
matplotlib-inline==0.1.7
mpmath==1.3.0
msrest==0.7.1
munkres==1.1.4
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.3
nfoursid==1.0.1
nltk==3.6.7
numba==0.59.1
numpy==1.26.4
oauthlib==3.2.2
openapi-codec==1.3.2
opencv-contrib-python==4.10.0.82
packaging==23.0
pandas==1.3.5
parso==0.8.4
patsy==0.5.6
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
plotly==5.22.0
pmdarima==2.0.4
prompt_toolkit==3.0.47
prophet==1.1.5
proto-plus==1.22.2
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.1
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.18.0
PyMeeus==0.5.12
pyod==2.0.1
pyparsing==3.1.2
pyRFC3339==1.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.1
python-json-logger==2.0.4
pytorch-lightning==2.3.3
pytz==2024.1
pywin32==306
PyYAML==6.0
pyzmq==26.0.3
redis==4.4.2
regex==2022.10.31
requests==2.32.3
requests-oauthlib==1.3.1
rq==1.10.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.3.7
scikit-learn==1.5.1
scipy==1.14.0
semver==2.13.0
sentry-sdk==1.14.0
setuptools==69.5.1
shap==0.45.1
shortuuid==1.0.13
simplejson==3.18.1
six==1.16.0
slicer==0.0.8
sqlparse==0.4.3
stack-data==0.6.2
stanio==0.5.1
statsforecast==1.7.5
statsmodels==0.14.2
sympy==1.13.0
tbats==1.1.3
tenacity==8.5.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
torch==2.3.1
torchmetrics==1.4.0.post0
tornado==6.4.1
tqdm==4.64.1
traitlets==5.14.3
triad==0.9.8
typing_extensions==4.12.2
tzdata==2024.1
u8darts==0.30.0
ujson==5.7.0
unicodedata2==15.1.0
uritemplate==4.1.1
urllib3==2.2.2
user-agents==2.2.0
utilsforecast==0.1.12
wcwidth==0.2.13
wheel==0.43.0
win-inet-pton==1.1.0
xarray==2024.6.0
xgboost==2.0.3
zipp==3.19.2
zstandard==0.23.0
